QComboBox doesn't have an arrow, so draw a line instead.
Unfortunately, Qt stylesheets don't allow for inline images via base64
and png or svg. Additionally, I have no interest in bundling any kind of
resource with this pack. Prior to this adjustment, a QComboBox appeared
roughly similar to a QLabel. With it, a line is drawn to the left of
where an arrow would be. This is not great, but hopefully sufficient
in terms of making it clearer that a QComboBox is, in fact, a QComboBox.